[FE] refactor : API Export 방식 통일
