+++ frontend/src/api/index.ts
@@ -1,2 +1,4 @@
+export * as authAPI from './auth';
 export * as reviewAPI from './review';
 export * as userAPI from './user';
+export * as templateAPI from './template';
